Fixes for sleep functionality
Rigidbody's sleep functionality now works independent of framerate. sleepFrameCount variable was replaced by inactiveTime variable, which tracks the time in seconds the object has been inactive for.
Added a new function QualifiedForSleep which checks whether or not the object is considered 'inactive'. 
Awake checking in ResolveCollision function now first checks if one object involved is awake and the other is not. The function then checks if at least one of the objects are not considered inactive (they are speeding up, moving too fast, etc) before resetting sleep variables for each object.
angularVelocityPrevious variable was added to keep track of the object's velocity during the previous frame. This is used to check whether or not the object's torque has increased.
Low-values for sleep qualification are now stored as static const variables, rather than being hard-coded into the function.
Removed obsolete lines of code in Physics and Box cpp files